Move split file
