+++ snuba/datasets/errors.py
@@ -38,7 +38,7 @@ from snuba.query.project_extension import ProjectExtension, ProjectWithGroupsPro
 from snuba.query.query import Query
 from snuba.query.query_processor import QueryProcessor
 from snuba.query.timeseries import TimeSeriesExtension
-from snuba.web.split import (
+from snuba.datasets.plans.split import (
     ColumnSplitSpec,
     SplitQueryPlanExecutionStrategy,
 )
